little rework
ive been doing alot of the work between pages in the user and in the home controllers. After a bit of thinking I realised it was getting a little too much and added an admin controller. Here i will put everything instructors and admin accounts need to handle. I also decided that javascript calling through the api controller was getting too complicated to keep track of and I thus have started using the default CRUD method. This is gonna be confusing later on but I got this