panic finished
as the hour of the presentation comes closer, I must finish what I have not done yet. I will push through and work out as much as I can in this little amount of time that I have. This commit message will contain nothing but strenght and courage to push through even in the face of fear. You can do this. This is absolutely worth it...